refactor: moved save button out of the center toolbar
